cureCuma/requirements.txt
# flask / web
flask
flask_bcrypt
requests
gunicorn

# db 
SQLAlchemy
pymysql

# test
pytest
black
pylint
autopep8

# misc
IPython
python-dateutil
